moves linting to worker thread during build
