+++ packages/next/server/on-demand-entry-handler.ts
@@ -4,7 +4,7 @@ import { join, posix } from 'path'
 import { parse } from 'url'
 import { webpack } from 'next/dist/compiled/webpack/webpack'
 import * as Log from '../build/output/log'
-import { verifyEslintSetup } from '../lib/verifyEslintSetup'
+import { verifyAndLint } from '../lib/verifyAndLint'
 import {
   normalizePagePath,
   normalizePathSep,
@@ -196,7 +196,13 @@ export default function onDemandEntryHandler(
         }
 
         // TODO: Move out of hot-reloader into a separate process
-        if (eslint) verifyEslintSetup(process.cwd(), pagesDir, pagePath)
+        if (eslint) {
+          verifyAndLint(process.cwd(), pagesDir, pagePath).then(
+            ({ results, hasMessages }) => {
+              if (hasMessages) console.log(results)
+            }
+          )
+        }
 
         Log.event(`build page: ${normalizedPage}`)
